pyproject.toml
[project]
name = "universitly"
version = "0.1.0"
description = ""
authors = [
    {name = "defuzia",email = "zhyvhyn.roman@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "phonenumbers (>=8.13.54,<9.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "celery (>=5.4.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-celery-results (>=2.5.1,<3.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
